feat : add frontend and fixed bugs in the backend

+++ src/main/java/com/team12/parkinglotsystem/strategy/FeeStrategy.java
@@ -1,5 +1,5 @@
 package com.team12.parkinglotsystem.strategy;
 
 public interface FeeStrategy {
-    double calculateFee(int hours);
+    double calculateFee(double hours);
 }
